+++ src/Bbob/Cli/CliShared.cs
@@ -1,3 +1,5 @@
+using System.Net.NetworkInformation;
+
 namespace Bbob.Main.Cli;
 
 public static class CliShared
@@ -23,5 +25,20 @@ public static class CliShared
         }
 
         return TextType.None;
+    }
+
+    public static int GetAvailablePort(int startingPort)
+    {
+        if (startingPort > ushort.MaxValue) throw new ArgumentException($"Can't be greater than {ushort.MaxValue}", nameof(startingPort));
+        var ipGlobalProperties = IPGlobalProperties.GetIPGlobalProperties();
+
+        var connectionsEndpoints = ipGlobalProperties.GetActiveTcpConnections().Select(c => c.LocalEndPoint);
+        var tcpListenersEndpoints = ipGlobalProperties.GetActiveTcpListeners();
+        var udpListenersEndpoints = ipGlobalProperties.GetActiveUdpListeners();
+        var portsInUse = connectionsEndpoints.Concat(tcpListenersEndpoints)
+                                             .Concat(udpListenersEndpoints)
+                                             .Select(e => e.Port);
+
+        return Enumerable.Range(startingPort, ushort.MaxValue - startingPort + 1).Except(portsInUse).FirstOrDefault();
     }
 }